Moved many inclusions from .h's to .cpp's
Moving the inclusions to the .cpp files when they aren't needed in the
.h files fixed the circular dependency issue we ran into earlier and
allowed me to change the subsystems back to their type in the
declarations and remove the static_cast stuff. Also added
std::make_shared<> for the robot subsystems and the OI

+++ src/Subsystems/DriveTrain.cpp
@@ -1,4 +1,6 @@
 #include "DriveTrain.h"
+#include "../Commands/DriveWithController.h"
+#include "../RobotMap.h"
 
 DriveTrain::DriveTrain() : Subsystem("DriveTrain") {
 	x = 0;
@@ -72,7 +74,7 @@ void DriveTrain::TurnToDegree(double angle) {
     currentAngle = gyro->GetAngle();
 }
 
-void DriveTrain::PIDWrite(float output) {
+void DriveTrain::PIDWrite(double output) {
 	rotateToAngleRate = output;
 }
